added load balance weight function to the tcpClient. Once a module register to the RC center, it will tell the RC the current weight it has . and when the client get the module instance list, it will get the weight info as well

+++ xixiRouter/src/main/java/xixi/router/Router.java
@@ -16,7 +16,7 @@ public interface Router{
 	
 	void removeTcpClient(String ipAddress);
 	
-	int moduleId();
+	short moduleId();
 	
 	public void router(RpcMessage message);
